Use only equality for title assertion
